+++ changelog.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [unreleased]
 
+- Add optional preface to usage output
 - Remove method Parser.WriteUsageTo, use Parser.Usage().WriteTo
 - Add type Usage